requirements.txt
fastapi
uvicorn[standard]
langchain
pytest
httpx
locust
pydantic
torch
transformers
requests

python-dotenv
faiss-cpu
sentence-transformers
langdetect
nltk